CPT-aanpassingen in eigen bestand
